+++ Assets/Scripts/UnitObject.cs
@@ -2,9 +2,9 @@ public class UnitObject
 {
     public bool Enabled { get; set; }
     public string Name { get; set; }
-    public int Level { get; set; }
-    public float Price;
-    public float PriceFactor { get; set; }
+    public long Level { get; set; }
+    public double Price;
+    public double PriceFactor { get; set; }
     public float PollutionClean { get; set; }
     public float PollutionCleanFactor { get; set; }
     public float PollutionUnlockPercentage;
